Fix LC_ID_DYLIB for OS X
Wrapping the SHLINKFLAGS in an array rather than as a single string
caused the link flags to be passed to clang double quoted. This cuased
clang to interpret the flags as a string rather than as a flag which
caused the -install_name flag to never be passed to the linker.